mcp-server/requirements.txt
mcp
pandas
python-dotenv
requests
google-generativeai
